refactor(http): Update return styles and imports
